MakeDepend now keeps files marked private out of the library
My previous change to throw everything in the library also included
LinkTimestamp.o which is the one file that you wouldn't want there.

MakeDepend already recognized a "private" tag, i.e.:

// MakeDepend: private

but this wasn't used anywhere.  Now it is -- if a file is marked
as private AND is not a dependency of anything in the repository,
then it won't be built into the library